requirements.txt
torch
pennylane<=0.33
scikit-learn
numpy<=1.23
matplotlib
streamlit
Pillow